# backend/contest_training_system/docs/api.yml
openapi: 3.0.0
info:
  title: 竞赛集训系统 API
  description: '根据提供的Markdown文档生成的完整版竞赛集训系统RESTful API'
  version: '1.0.0'
servers:
  - url: https://api.example.com
    description: '生产服务器'
...
paths:
...
  /problem/create:
    post:
      tags:
        - '3. 题库管理'
      summary: '3.1 新建题目'
      description: '教师新建题目'
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - inputSpec
                - outputSpec
                - sampleInput
                - sampleOutput
                - timeLimit
                - memoryLimit
                - testdataZip
              properties:
                title:
                  type: string
                  example: "A+B Problem"
                description:
                  type: string
                  example: "计算两个数之和"
                inputSpec:
                  type: string
                  example: "输入两个整数"
                outputSpec:
                  type: string
                  example: "输出一个整数"
                sampleInput:
                  type: string
                  example: "1 2"
                sampleOutput:
                  type: string
                  example: "3"
                remark:
                  type: string
                  example: "无"
                timeLimit:
                  type: integer
                  example: 1000
                memoryLimit:
                  type: integer
                  example: 256
                testdataZip:
                  type: string
                  example: "url/to/zip"
      responses:
        '200':
          description: '创建成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 1
                  msg:
                    type: string
                    example: '创建成功'
                  data:
                    type: object
                    properties:
                      problemId:
                        type: integer
                        example: 201
  /problem/update:
    put:
      tags:
        - '3. 题库管理'
      summary: '3.2 修改题目'
      description: '教师修改题目信息'
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - problemId
              properties:
                problemId:
                  type: integer
                  example: 201
                title:
                  type: string
                  example: "A+B Problem（修改版）"
                description:
                  type: string
                  example: "计算两个数之和（修改版）"
                inputSpec:
                  type: string
                outputSpec:
                  type: string
                sampleInput:
                  type: string
                sampleOutput:
                  type: string
                remark:
                  type: string
                timeLimit:
                  type: integer
                  example: 2000
                memoryLimit:
                  type: integer
      responses:
        '200':
          description: '修改成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 1
                  msg:
                    type: string
                    example: "修改成功"
                  data:
                    type: object
                    properties:
                      problemId:
                        type: integer
                        example: 201
                      title:
                        type: string
                        example: "A+B Problem（修改版）"
  /problem/{problemId}:
    get:
      tags:
        - '3. 题库管理'
      summary: '3.5 获取题目详情'
      description: '获取题目详细信息'
      security:
        - bearerAuth: []
      parameters:
        - name: problemId
          in: path
          required: true
          schema:
            type: integer
          example: 201
      responses:
        '200':
          description: '获取成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 1
                  msg:
                    type: string
                    example: 'success'
                  data:
                    type: object
                    properties:
                      problemId:
                        type: integer
                      title:
                        type: string
                      description:
                        type: string
                      inputSpec:
                        type: string
                      outputSpec:
                        type: string
                      sampleInput:
                        type: string
                      sampleOutput:
                        type: string
                      remark:
                        type: string
                      timeLimit:
                        type: integer
                      memoryLimit:
                        type: integer
                      creatorId:
                        type: integer
                      isHidden:
                        type: integer
                      createdAt:
                        type: string
                        format: date-time
                      datasets:
                        type: array
                        items:
                          type: object
                          properties:
                            datasetId:
                              type: integer
                            version:
                              type: integer
                            zipUrl:
                              type: string
                            isActive:
                              type: integer
                            addedAt:
                              type: string
                              format: date-time
    delete:
      tags:
        - '3. 题库管理'
      summary: '3.3 删除/隐藏题目'
      description: '删除或隐藏题目（未在赛事中使用可删除，否则隐藏）'
      security:
        - bearerAuth: []
      parameters:
        - name: problemId
          in: path
          required: true
          schema:
            type: integer
          example: 201
      responses:
        '200':
          description: '操作成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 1
                  msg:
                    type: string
                    example: "题目已隐藏（在赛事中使用）"
                  data:
                    type: object
                    properties:
                      problemId:
                        type: integer
                        example: 201
                      action:
                        type: string
                        example: "hidden"
  /problem/list:
    get:
      tags:
        - '3. 题库管理'
      summary: '3.4 查询题目列表'
      description: '查询题目列表，支持分页、筛选和排序'
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: keyword
          in: query
          schema:
            type: string
        - name: creatorId
          in: query
          schema:
            type: integer
        - name: isHidden
          in: query
          schema:
            type: integer
            enum: [0, 1]
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [id_asc, id_desc, title_asc, title_desc, created_asc, created_desc]
            default: 'id_asc'
      responses:
        '200':
          description: '查询成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 1
                  msg:
                    type: string
                    example: 'success'
                  data:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 50
                      page:
                        type: integer
                        example: 1
                      pageSize:
                        type: integer
                        example: 10
                      problems:
                        type: array
                        items:
                          type: object
                          properties:
                            problemId:
                              type: integer
                            title:
                              type: string
                            timeLimit:
                              type: integer
                            memoryLimit:
                              type: integer
                            creatorId:
                              type: integer
                            isHidden:
                              type: integer
                            createdAt:
                              type: string
                              format: date-time
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '在需要鉴权的接口的请求头中添加 "Authorization: Bearer <YOUR_TOKEN>"'